server/requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
pymysql==1.1.0
cryptography==44.0.3
python-dotenv==1.0.0
requests==2.31.0
shopifyapi==12.4.0
google-generativeai==0.3.1
gunicorn==21.2.0